requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
boto3==1.28.38
